refactor(gathersg): extract shared case-fields schema
create-case and update-case duplicated the exact same caseFields
parsing/validation logic. Extract it into a single caseFieldsSchema,
and move fieldTypeEnum into common/constants.ts as its single source
of truth instead of being declared separately in each schema file.

Also adds an inert `fieldType === 'email'` branch (email validation
via zod). It can't be reached yet since fieldTypeEnum doesn't include
'email', but keeps this refactor and the upcoming email field type
change as separate, reviewable steps.
